Moved Profet model logic to model_utils.py
